+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import checkr
+#' @import chk
 #' @importFrom stats reshape
 #' @importFrom utils download.file read.csv
 NULL