+++ NEWS.md
@@ -1,3 +1,5 @@
+- Replaced checkr with chk.
+
 # rpdo 0.2.5
 
 - Added values from July 2018 to September 2018.